refactor: Dealer 내 불필요한 책임을 가지는 메소드 제거
```
딜러가 너무 많은 책임을 가지고 있는 것 같아요.

게임 상태를 계산하는 부분은 Players에서 진행해주는게 좋을 것 같습니다.

_Originally posted by @hyunniiii in https://github.com/next-step/kotlin-blackjack/pull/354#discussion_r1048261145_

```
